Refactor PricingTransferRequest component; remove custom dropdown logic and replace with GenerateActionButton component for improved action handling
