+++ src/components/organisms/logistics/orders/transfer_request/PricingTransferRequest.scss
@@ -649,29 +649,3 @@
         cursor: not-allowed;
     }
 }
-
-.dropdownGenerateAction {
-    ul {
-        background-color: $white !important;
-        padding: 12px !important;
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-    }
-
-    li {
-        background-color: $background !important;
-        border: 1px solid transparent;
-
-        &:hover {
-            border: 1px solid $black !important;
-        }
-    }
-}
-
-.generateActionBtn {
-    height: 48px !important;
-    font-size: 1rem !important;
-    font-weight: 600 !important;
-    box-shadow: none !important;
-}
